+++ src/features/events/providers/event-position-provider.tsx
@@ -6,6 +6,7 @@ import { createContextWithUseHook } from '@/util/create-context-with-use-hook';
 import { useStaleTimestamp } from '@/hooks/use-stale-timestamp';
 import { useEventContext } from './event-provider';
 import { socket } from '@/socket';
+import { useSessionStorage } from '@/hooks/use-session-storage';
 
 const [EventPositionContext, useEventPositionContext] = createContextWithUseHook<{
   position: TEvent['position'];
@@ -18,7 +19,11 @@ type EventPositionProviderProps = React.PropsWithChildren;
 export function EventPositionProvider({ children }: EventPositionProviderProps) {
   const { event } = useEventContext();
 
-  const [position, setPosition] = useState(event?.position);
+  const [position, setPosition] = useSessionStorage(
+    ['conjurence-event-position', event?.id],
+    event?.position,
+    true
+  );
 
   //Invalidate positions of mobile events if they haven't been updated for a while.
   const positionIsStale = useStaleTimestamp(