+++ src/hooks/use-infinite-scroll.ts
@@ -1,6 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
 import { useInView } from 'react-intersection-observer';
 
+/**@todo */
 export function useInfiniteScroll(initialPage: number = 0, queryFn: () => Promise<TODO>) {
   const [page, setPage] = useState(initialPage);
   const { inView, ref: observedRef } = useInView();