+++ src/hooks/use-session-storage.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 
 function getValue<T>(key: string, initialValue?: T): T {
   if (typeof sessionStorage === 'undefined') {
@@ -9,13 +9,20 @@ function getValue<T>(key: string, initialValue?: T): T {
   return value ? JSON.parse(value) : initialValue;
 }
 
-export function useSessionStorage<T>(key: string, initialValue?: T) {
-  const [value, setValue] = useState(() => getValue(key, initialValue));
+export function useSessionStorage<T>(key: string[], initialValue?: T, cleanupOnUnmount?: boolean) {
+  const combinedKey = useMemo(() => key.join('-'), [key]);
+  const [value, setValue] = useState(() => getValue(combinedKey, initialValue));
 
   useEffect(() => {
     if (typeof sessionStorage === 'undefined') return;
-    sessionStorage.setItem(key, typeof value !== 'string' ? JSON.stringify(value) : value);
-  }, [value, key]);
+    sessionStorage.setItem(combinedKey, typeof value !== 'string' ? JSON.stringify(value) : value);
+
+    return () => {
+      if (cleanupOnUnmount) {
+        sessionStorage.removeItem(combinedKey);
+      }
+    };
+  }, [value, combinedKey]);
 
   return [value, setValue] as const;
 }